refactor(frontend): split InvestmentAnalyticsPage and healthScoreUtils under 500 LOC
InvestmentAnalyticsPage 677 -> 88 LOC orchestrator with 4 sub-components.
healthScoreUtils 667 -> 27 LOC barrel; split into types/analysis/scorers.
